+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Release (2025-xx-xx)
+- `stackitmarketplace`: [v1.10.0](services/stackitmarketplace/CHANGELOG.md#v1100)
+  - **Feature:** Added `PlanId` to `CatalogProductPricingOption`and `SubscriptionProduct`
+
 ## Release (2025-09-11)
 - `cdn`: [v1.6.0](services/cdn/CHANGELOG.md#v160)
   - **Feature:** Added Attribute `LogSink` to `ConfigPatch`